Revert "refactor(engine): extract KVCacheInitializer (init_kv_cache) to its own TU (#335)"
This reverts commit e44cbcf95a3f5ce822cf89946fd1900aee53e1e3.
