+++ CMakeLists.txt
@@ -221,8 +221,6 @@ set(IMP_RUNTIME_SOURCES
     src/runtime/batch.cpp
     src/runtime/engine.cpp
     src/runtime/engine_init_resolver.cpp
-    src/runtime/engine_weight_upload.cpp
-    src/runtime/engine_kv_cache_init.cpp
     src/runtime/engine_graph_decode.cpp
     src/runtime/vision_pipeline.cpp
     src/runtime/constraint_manager.cpp
