+++ src/runtime/engine.cpp
@@ -612,6 +612,492 @@ bool Engine::init(std::shared_ptr<Model> model, const EngineConfig& config) {
     return true;
 }
 
+bool Engine::init_weights() {
+    const auto& mcfg = model_->config();
+
+    // Initialize graph executor (Phase 1: compute sizes, no GPU allocation)
+    executor_ = std::make_unique<GraphExecutor>();
+    executor_->set_vram_allocator(&memory_manager_.vram_allocator());
+    {
+        int eff_batch = config_.max_batch_size;
+        if (!executor_->init(*model_, config_.compute_dtype, config_.use_pdl, eff_batch, config_.max_seq_len,
+                             config_.use_fp8_prefill, config_.use_nvfp4_decode, config_.use_mxfp4_prefill))
+            return false;
+
+        if (config_.dual_path_quant) {
+            executor_->set_dual_path_quant(true);
+            IMP_LOG_INFO("Dual-path quant: attention weights → FP8, FFN weights → NVFP4");
+        }
+
+        if (config_.streaming_kv_enabled) {
+            // Streaming is only safe for the FP16 GQA decode kernel — the
+            // quantized variants don't yet skip -1 sentinels in their block
+            // tables. Refuse to enable streaming with non-FP16 KV caches so
+            // we never call evict_middle_blocks for an unsupported path.
+            if (config_.kv_cache_dtype != QType::F16) {
+                IMP_LOG_WARN(
+                    "StreamingLLM smart KV cache requires FP16 KV cache "
+                    "(requested %d) — disabling streaming.",
+                    static_cast<int>(config_.kv_cache_dtype));
+                config_.streaming_kv_enabled = false;
+            } else {
+                int n_sinks = (config_.streaming_kv_n_sinks > 0) ? config_.streaming_kv_n_sinks : 4;
+                int win = (config_.streaming_kv_window > 0) ? config_.streaming_kv_window
+                                                            : model_->config().sliding_window;
+                executor_->set_streaming_kv(n_sinks, win);
+                if (n_sinks > 0 && win > 0) {
+                    IMP_LOG_INFO("StreamingLLM smart KV cache enabled: %d sinks + %d-token window", n_sinks,
+                                 win);
+                    // Block-table contents change every step once eviction
+                    // begins; a CUDA graph captured against an old table
+                    // would replay stale pointers. Re-capturing per step
+                    // negates the graph's win, so disable graphs entirely.
+                    if (config_.use_cuda_graphs) {
+                        IMP_LOG_INFO(
+                            "Disabling CUDA Graphs while StreamingLLM is active "
+                            "(block table mutates per decode step).");
+                        config_.use_cuda_graphs = false;
+                    }
+                } else {
+                    IMP_LOG_WARN(
+                        "StreamingLLM enabled but no sliding window configured "
+                        "(n_sinks=%d, window=%d) — disabling streaming.",
+                        n_sinks, win);
+                    config_.streaming_kv_enabled = false;
+                }
+            }
+        }
+    }
+
+    // Reserve L2 persisting cache for decode GEMV
+    {
+        cudaDeviceProp prop;
+        cudaGetDeviceProperties(&prop, 0);
+        size_t max_persist = prop.persistingL2CacheMaxSize;
+        if (max_persist > 0) {
+            size_t reserve = max_persist * 3 / 4;
+            cudaDeviceSetLimit(cudaLimitPersistingL2CacheSize, reserve);
+            IMP_LOG_INFO("L2 persisting cache: reserved %zu MB / %zu MB total", reserve >> 20,
+                         max_persist >> 20);
+        }
+    }
+
+    // Tune the default cudaMallocAsync pool so it retains freed memory instead
+    // of returning it to the driver. Many paths (prefill metadata, MoE scratch,
+    // spec decoder block tables, vision staging) use cudaMallocAsync with the
+    // default pool; the default threshold is 0, which calls cuMemUnmap on every
+    // free. Setting UINT64_MAX keeps allocations for re-use — the KV cache and
+    // workspaces already own their memory through the DeviceAllocator, so the
+    // default-pool footprint stays small.
+    {
+        cudaMemPool_t default_pool = nullptr;
+        int dev = 0;
+        cudaGetDevice(&dev);
+        if (cudaDeviceGetDefaultMemPool(&default_pool, dev) == cudaSuccess && default_pool != nullptr) {
+            uint64_t threshold = UINT64_MAX;
+            cudaMemPoolSetAttribute(default_pool, cudaMemPoolAttrReleaseThreshold, &threshold);
+        }
+    }
+
+    // Compute VRAM reserve for expert weight upload
+    size_t expert_reserve = executor_->workspace_estimate();
+    {
+        int head_dim_est = mcfg.head_dim > 0 ? mcfg.head_dim : (mcfg.d_model / mcfg.n_heads);
+        size_t elem_sz = dtype_size(config_.kv_cache_dtype);
+        int est_bs = config_.kv_block_size > 0 ? config_.kv_block_size : kKVBlockSize;
+        int blocks_per_seq = (config_.max_seq_len + est_bs - 1) / est_bs;
+        int n_attn = 0;
+        for (int i = 0; i < mcfg.n_layers; i++)
+            if (model_->layer(i).wq.data != nullptr)
+                n_attn++;
+        if (n_attn == 0)
+            n_attn = mcfg.n_layers;
+        size_t kv_block_bytes = static_cast<size_t>(est_bs) * mcfg.n_kv_heads * head_dim_est * elem_sz;
+        size_t kv_est = static_cast<size_t>(blocks_per_seq * config_.max_batch_size) * 2 * n_attn *
+                        kv_block_bytes;
+        {
+            size_t total_vram = 0, f = 0;
+            cudaMemGetInfo(&f, &total_vram);
+            // For large MoE models (128 experts), prefer fitting all experts on GPU
+            // over reserving huge KV cache. All-GPU experts enable the decode fast
+            // path (dp4a GEMV, no D2H sync) and CUDA graph capture.
+            size_t vram_frac = (mcfg.n_experts > 16) ? 10 : 5;
+            kv_est = std::min(kv_est, total_vram / vram_frac);
+        }
+        expert_reserve += kv_est;
+
+        if (mcfg.ssm_inner_size > 0) {
+            int conv_ch = mcfg.ssm_inner_size + 2 * mcfg.ssm_group_count * mcfg.ssm_state_size;
+            int n_heads = mcfg.ssm_dt_rank;
+            int hd_ssm = (n_heads > 0) ? mcfg.ssm_inner_size / n_heads : 0;
+            int n_ssm = 0;
+            for (int i = 0; i < mcfg.n_layers; i++)
+                if (model_->layer(i).ssm_in.data != nullptr)
+                    n_ssm++;
+            expert_reserve += static_cast<size_t>(n_ssm) * config_.max_batch_size *
+                              (conv_ch * std::max(mcfg.ssm_conv_kernel - 1, 0) * sizeof(float) +
+                               n_heads * hd_ssm * mcfg.ssm_state_size * dtype_size(config_.ssm_state_dtype));
+        }
+
+        size_t safety = 256ULL * 1024 * 1024;  // base safety
+        // Only add safety for features that will actually allocate VRAM.
+        // On tight VRAM models (Nemotron-30B), every MiB matters for expert coverage.
+        expert_reserve += safety;
+
+        IMP_LOG_INFO("Expert upload reserve: %.2f MiB (workspace=%.2f, kv=%.2f, ssm+safety=rest)",
+                     expert_reserve / (1024.0 * 1024.0), executor_->workspace_estimate() / (1024.0 * 1024.0),
+                     kv_est / (1024.0 * 1024.0));
+    }
+
+    // Upload weights
+    size_t free_before = 0, total_before = 0;
+    cudaMemGetInfo(&free_before, &total_before);
+    IMP_LOG_INFO("GPU memory before weight upload: %zu MiB free / %zu MiB total", free_before / (1024 * 1024),
+                 total_before / (1024 * 1024));
+
+    cudaStream_t upload_stream = nullptr;
+    IMP_CUDA_CHECK_LOG(cudaStreamCreateWithFlags(&upload_stream, cudaStreamNonBlocking));
+
+    if (!model_->upload_weights_gpu(config_.compute_dtype, upload_stream ? upload_stream : stream_,
+                                    expert_reserve)) {
+        IMP_LOG_ERROR("Weight upload failed. Try a smaller quantization.");
+        if (upload_stream)
+            IMP_CUDA_CHECK_LOG(cudaStreamDestroy(upload_stream));
+        return false;
+    }
+
+    if (upload_stream) {
+        cudaEvent_t upload_done;
+        IMP_CUDA_CHECK_LOG(cudaEventCreate(&upload_done));
+        IMP_CUDA_CHECK_LOG(cudaEventRecord(upload_done, upload_stream));
+        IMP_CUDA_CHECK_LOG(cudaStreamWaitEvent(stream_, upload_done));
+        IMP_CUDA_CHECK_LOG(cudaEventDestroy(upload_done));
+        IMP_CUDA_CHECK_LOG(cudaStreamDestroy(upload_stream));
+    }
+
+    size_t free_after = 0, total_after = 0;
+    cudaMemGetInfo(&free_after, &total_after);
+    IMP_LOG_INFO("GPU memory after weight upload: %zu MiB free / %zu MiB total (weights ~%zu MiB)",
+                 free_after / (1024 * 1024), total_after / (1024 * 1024),
+                 (free_before - free_after) / (1024 * 1024));
+
+    // Check for host-resident expert weights
+    if (mcfg.n_experts > 0) {
+        for (int i = 0; i < mcfg.n_layers; i++) {
+            if (model_->layer(i).expert_up_packed.data && !model_->layer(i).expert_up_packed.on_device) {
+                experts_on_host_ = true;
+                break;
+            }
+        }
+        if (experts_on_host_ && config_.use_cuda_graphs) {
+            // Phase 5: the opt-in `moe.allow_graphs_under_offload` flag skips
+            // this guard so the user can experiment with captured decode
+            // under host-offload. Correctness is conditional on prefetch
+            // coverage matching router selection — captured cudaMemcpyAsync
+            // nodes have fixed (src host ptr, dst slot) pairs that don't
+            // adapt to per-token routing changes. See config.h doc for the
+            // architectural caveat; Phase 5.1+ refactors dispatch kernels
+            // to read the device mirror at runtime so the captured graph
+            // adapts correctly.
+            if (RuntimeConfig::current().moe.allow_graphs_under_offload) {
+                IMP_LOG_WARN(
+                    "CUDA graphs ENABLED under host-offload "
+                    "(moe.allow_graphs_under_offload=true). EXPERIMENTAL: output "
+                    "correctness depends on prefetch coverage matching router "
+                    "selection. Set moe.prefetch_top_k high enough that the "
+                    "captured top-K covers the router's hot set.");
+            } else {
+                IMP_LOG_INFO("Disabling CUDA graphs: expert weights on host");
+                IMP_LOG_INFO(
+                    "  Tip: if model+KV fits in VRAM, set IMP_EXPERT_OVERHEAD_PCT=10 "
+                    "(default 30) to upload ALL experts and re-enable CUDA graphs "
+                    "(+~180%% decode on Qwen 3.6 35B Q4_K_M).");
+                IMP_LOG_INFO(
+                    "  Or set moe.allow_graphs_under_offload=true (experimental) "
+                    "to opt into captured decode under host-offload.");
+                config_.use_cuda_graphs = false;
+            }
+        }
+        if (RuntimeConfig::current().runtime.cuda_graphs == "never" && config_.use_cuda_graphs) {
+            IMP_LOG_INFO("Disabling CUDA graphs: runtime.cuda_graphs=never");
+            config_.use_cuda_graphs = false;
+        }
+        // MoE decode fast path is fully device-side (no D2H memcpy) — graph-safe.
+        // Only MoE prefill paths use D2H sync for expert_offsets, but prefill is
+        // never captured in CUDA graphs.
+    }
+
+    // Phase 2: allocate GPU workspace
+    (void)executor_->allocate_workspaces(experts_on_host_);
+
+    // Layer offloading
+    if (config_.gpu_layers >= 0) {
+        offload_mgr_ = std::make_unique<LayerOffloadManager>();
+        if (!offload_mgr_->init(model_.get(), config_.gpu_layers)) {
+            IMP_LOG_WARN("Layer offloading init failed, continuing without it");
+            offload_mgr_.reset();
+        }
+    }
+
+    return true;
+}
+
+bool Engine::init_kv_cache() {
+    const auto& mcfg = model_->config();
+    int head_dim = mcfg.head_dim > 0 ? mcfg.head_dim : (mcfg.d_model / mcfg.n_heads);
+
+    // Build KV layer mapping for hybrid models
+    int n_attn_layers = 0;
+    std::vector<int> kv_layer_map(mcfg.n_layers, -1);
+    for (int i = 0; i < mcfg.n_layers; i++) {
+        if (model_->layer(i).wq.data != nullptr && model_->layer(i).gdn_gate.data == nullptr)
+            kv_layer_map[i] = n_attn_layers++;
+    }
+    if (n_attn_layers == 0) {
+        n_attn_layers = mcfg.n_layers;
+        for (int i = 0; i < mcfg.n_layers; i++)
+            kv_layer_map[i] = i;
+    }
+    int n_kv_layers = n_attn_layers;
+    IMP_LOG_INFO("KV cache layers: %d attention out of %d total", n_kv_layers, mcfg.n_layers);
+
+    // Auto-select block size
+    if (config_.kv_block_size <= 0) {
+        config_.kv_block_size = (mcfg.n_kv_heads <= 4 && mcfg.n_kv_heads > 0) ? 32 : kKVBlockSize;
+        IMP_LOG_INFO("KV block size: auto → %d (n_kv_heads=%d)", config_.kv_block_size, mcfg.n_kv_heads);
+    }
+    const int kv_bs = config_.kv_block_size;
+    int blocks_per_seq = (config_.max_seq_len + kv_bs - 1) / kv_bs;
+
+    // VRAM budget
+    auto vram_budget =
+        memory_manager_.compute_budget(*model_, config_, n_kv_layers, head_dim, effective_free_vram());
+    int max_blocks = config_.kv_cache_max_blocks > 0 ? config_.kv_cache_max_blocks
+                                                     : vram_budget.kv_max_blocks;
+
+    {
+        QType kv_dtype = config_.kv_cache_dtype;
+        size_t block_bytes = static_cast<size_t>(kv_bs) * mcfg.n_kv_heads * head_dim * dtype_size(kv_dtype);
+        size_t total_kv = static_cast<size_t>(n_kv_layers) * max_blocks * 2 * block_bytes;
+        IMP_LOG_INFO(
+            "KV cache: %d blocks (%.0f tokens), %.2f MiB, dtype=%s "
+            "(layers=%d/%d, kv_heads=%d, head_dim=%d, block_size=%d)",
+            max_blocks, static_cast<double>(max_blocks) * kv_bs,
+            static_cast<double>(total_kv) / (1024.0 * 1024.0), dtype_name(kv_dtype), n_kv_layers,
+            mcfg.n_layers, mcfg.n_kv_heads, head_dim, kv_bs);
+    }
+
+    // Per-layer KV shape path (Gemma 4 dual attention geometry): build per-layer
+    // nkv/hd arrays restricted to attention layers (hybrid models may have non-attn layers).
+    std::unique_ptr<KVCache> kv_cache;
+    if (!mcfg.head_dim_per_layer.empty() && config_.kv_cache_dtype != QType::INT8 &&
+        config_.kv_cache_dtype != QType::INT4) {
+        std::vector<int> per_layer_nkv(n_kv_layers, 0);
+        std::vector<int> per_layer_hd(n_kv_layers, 0);
+        for (int l = 0, k = 0; l < mcfg.n_layers && k < n_kv_layers; l++) {
+            // Only attention layers get KV cache entries
+            int attn_nkv = (l < (int)mcfg.n_kv_heads_per_layer.size()) ? mcfg.n_kv_heads_per_layer[l]
+                                                                       : mcfg.n_kv_heads;
+            if (attn_nkv <= 0)
+                continue;  // non-attention layer (SSM/GDN)
+            per_layer_nkv[k] = attn_nkv;
+            per_layer_hd[k] = (l < (int)mcfg.head_dim_per_layer.size() && mcfg.head_dim_per_layer[l] > 0)
+                                  ? mcfg.head_dim_per_layer[l]
+                                  : head_dim;
+            k++;
+        }
+        kv_cache = std::make_unique<KVCache>(n_kv_layers, per_layer_nkv, per_layer_hd, config_.kv_cache_dtype,
+                                             max_blocks, kv_bs, &memory_manager_.vram_allocator());
+    } else {
+        kv_cache = std::make_unique<KVCache>(n_kv_layers, mcfg.n_kv_heads, head_dim, config_.kv_cache_dtype,
+                                             max_blocks, kv_bs, &memory_manager_.vram_allocator());
+    }
+    kv_cache_raw_ = kv_cache.get();
+    kv_manager_ = std::make_unique<KVCacheManager>(std::move(kv_cache));
+
+    // BitDecoding Phase 3: residual FP16 cache (opt-in).
+    //
+    // Ring state (write_idx / fill_count per slot) lives in device memory
+    // (kv_manager_->d_residual_widx_ptr / d_residual_fc_ptr). Updated by a
+    // tiny advance_residual_state_kernel at the end of forward_logits; the
+    // residual write/read kernels read the state at execution time. This
+    // makes the whole path graph-capture-safe — graphs stay enabled.
+    {
+        const auto& rcfg = RuntimeConfig::current();
+        int residual_n = rcfg.kv_cache.bitdecoding_residual_tokens;
+        if (residual_n > 0 && config_.kv_cache_dtype == QType::NVFP4) {
+            int max_seqs = config_.max_batch_size > 0 ? config_.max_batch_size : 1;
+            if (kv_manager_->enable_residual_buffer(max_seqs, residual_n, &memory_manager_.vram_allocator())) {
+                // Persistent batch→slot lookup buffer (graph-safe). [max_batch_size] ints.
+                size_t slot_bytes = static_cast<size_t>(max_seqs) * sizeof(int);
+                cudaMalloc(&d_kv_slot_buf_, slot_bytes);
+                std::vector<int> init_slots(max_seqs, -1);
+                cudaMemcpy(d_kv_slot_buf_, init_slots.data(), slot_bytes, cudaMemcpyHostToDevice);
+                d_kv_slot_last_uploaded_.assign(max_seqs, -1);
+            }
+        } else if (residual_n > 0) {
+            IMP_LOG_INFO("kv_cache.bitdecoding_residual_tokens=%d ignored (only active with kv_cache_dtype=NVFP4)",
+                         residual_n);
+        }
+    }
+
+    if (config_.use_prefix_caching) {
+        if (mcfg.ssm_inner_size > 0) {
+            IMP_LOG_WARN(
+                "Prefix caching disabled for recurrent model — "
+                "SSM/GDN state requires full sequential prefill");
+        } else {
+            kv_manager_->set_prefix_caching_enabled(true);
+            IMP_LOG_INFO("Prefix caching enabled");
+            if (!config_.prefix_cache_path.empty()) {
+                int restored = kv_manager_->load_prefix_cache(config_.prefix_cache_path, stream_);
+                if (restored > 0)
+                    IMP_LOG_INFO("Restored %d prefix cache blocks from %s", restored,
+                                 config_.prefix_cache_path.c_str());
+            }
+        }
+    }
+
+    executor_->set_kv_layer_map(std::move(kv_layer_map));
+
+    if (offload_mgr_)
+        executor_->set_offload_manager(offload_mgr_.get());
+    scheduler_->set_kv_manager(kv_manager_.get());
+
+    // SSM state
+    if (mcfg.ssm_inner_size > 0) {
+        int n_ssm = 0;
+        for (int i = 0; i < mcfg.n_layers; i++)
+            if (model_->layer(i).ssm_in.data != nullptr)
+                n_ssm++;
+        if (n_ssm > 0) {
+            int conv_ch = mcfg.ssm_inner_size + 2 * mcfg.ssm_group_count * mcfg.ssm_state_size;
+            int n_heads = mcfg.ssm_dt_rank;
+            int hd = (n_heads > 0) ? mcfg.ssm_inner_size / n_heads : 0;
+            ssm_state_ = std::make_unique<SSMState>();
+            if (!ssm_state_->init(n_ssm, config_.max_batch_size, conv_ch, mcfg.ssm_conv_kernel, n_heads, hd,
+                                  mcfg.ssm_state_size, config_.ssm_state_dtype, &memory_manager_.vram_allocator())) {
+                IMP_LOG_WARN("Failed to init SSM state, continuing without it");
+                ssm_state_.reset();
+            }
+        }
+    }
+
+    // GDN detection
+    {
+        int n_gdn = 0;
+        for (int i = 0; i < mcfg.n_layers; i++)
+            if (model_->layer(i).gdn_gate.data != nullptr)
+                n_gdn++;
+        if (n_gdn > 0) {
+            if (config_.use_cuda_graphs) {
+                IMP_LOG_INFO("GDN model: %d layers, CUDA graphs enabled (recurrent state in-place)", n_gdn);
+            } else {
+                IMP_LOG_INFO(
+                    "GDN model: %d layers, CUDA graphs disabled (disabled earlier by caller or expert "
+                    "offload)",
+                    n_gdn);
+            }
+            // GDN recurrent state accumulates small precision errors per token.
+            // FP8 E4M3 (3-bit mantissa) amplifies these through the delta rule
+            // scan, causing degenerate output after ~50 special tokens in
+            // multi-turn chat.  Force FP16 weights for GDN prefill.
+            if (config_.use_fp8_prefill) {
+                if (config_.dual_path_quant) {
+                    IMP_LOG_WARN(
+                        "GDN + dual-path: attention weights forced to FP16 (not FP8) — "
+                        "recurrent state needs FP16 precision. FFN weights still use NVFP4.");
+                } else {
+                    IMP_LOG_INFO("GDN model: disabling FP8 prefill (recurrent state needs FP16 precision)");
+                }
+                config_.use_fp8_prefill = 0;
+                executor_->disable_fp8_prefill();
+            }
+        }
+    }
+
+    // (Gemma 4 FP8 prefill disabled earlier, before executor init)
+
+    // Detect pure Mamba2 SSM layers (layers with ssm_in but without gdn_gate).
+    // GDN-only models (Qwen3.5) are graph-compatible; pure SSM (Nemotron-H) is not yet.
+    {
+        int n_pure_ssm = 0;
+        for (int i = 0; i < mcfg.n_layers; i++)
+            if (model_->layer(i).ssm_in.data != nullptr && model_->layer(i).gdn_gate.data == nullptr)
+                n_pure_ssm++;
+        has_pure_ssm_layers_ = (n_pure_ssm > 0);
+    }
+
+    // Dequant weights → FP16/FP8/NVFP4 caches
+    executor_->pre_dequant_weights(stream_, vram_budget);
+    dequant_done_ = true;
+    cudaStreamSynchronize(stream_);
+
+    // Pre-allocate the gemm_nvfp4 fallback dequant workspace. Sized from
+    // wcache_.nvfp4 which is populated by pre_dequant_weights above, so this
+    // call must come AFTER. Lets the M>1 fallback path (used by future
+    // multi-token verify / spec-decode) run inside CUDA stream capture
+    // without crashing on cudaMalloc.
+    (void)executor_->allocate_nvfp4_dequant_workspace();
+    if (config_.use_fp8_prefill)
+        IMP_LOG_INFO("Weight cache: FP8 E4M3 (2x prefill throughput on sm_120)");
+
+    // Pre-allocate decode batch pool + penalty buffer
+    decode_batch_pool_.allocate(config_.max_batch_size, blocks_per_seq, &memory_manager_.vram_allocator());
+    {
+        d_penalty_tokens_capacity_ = static_cast<size_t>(config_.max_seq_len);
+        d_penalty_tokens_ = static_cast<int32_t*>(
+            memory_manager_.vram_allocator().allocate(d_penalty_tokens_capacity_ * sizeof(int32_t), "penalty_tokens"));
+        if (!d_penalty_tokens_) {
+            IMP_LOG_WARN("Failed to pre-allocate penalty token buffer");
+            d_penalty_tokens_capacity_ = 0;
+        }
+    }
+
+    // Pre-allocate prefill metadata pool (avoids per-request cudaMallocAsync)
+    {
+        size_t tok_bytes = config_.max_seq_len * sizeof(int32_t);
+        size_t pos_bytes = config_.max_seq_len * sizeof(int);
+        // A single request's block_table can grow to the entire KV cache
+        // pool (max_blocks), not just max_seq_len/block_size. Size from
+        // max_blocks so the H2D copy at the prefill metadata upload site
+        // doesn't overflow on long-cumulative-KV requests.
+        size_t bt_bytes = static_cast<size_t>(max_blocks) * sizeof(int);
+        size_t cl_bytes = sizeof(int);
+        prefill_pool_size_ = tok_bytes + pos_bytes + bt_bytes + cl_bytes;
+        prefill_pool_ = memory_manager_.vram_allocator().allocate(prefill_pool_size_, "prefill_pool");
+        if (prefill_pool_) {
+            auto* base = static_cast<char*>(prefill_pool_);
+            d_pf_token_ids_ = reinterpret_cast<int32_t*>(base);
+            d_pf_positions_ = reinterpret_cast<int*>(base + tok_bytes);
+            d_pf_block_tables_ = reinterpret_cast<int*>(base + tok_bytes + pos_bytes);
+            d_pf_context_lens_ = reinterpret_cast<int*>(base + tok_bytes + pos_bytes + bt_bytes);
+        } else {
+            IMP_LOG_WARN("Failed to pre-allocate prefill pool, will use per-request malloc");
+        }
+
+        // Pinned host staging buffers for prefill
+        if (cudaHostAlloc(&h_pf_positions_, config_.max_seq_len * sizeof(int), cudaHostAllocDefault) !=
+            cudaSuccess)
+            h_pf_positions_ = nullptr;
+        if (cudaHostAlloc(&h_pf_token_ids_, config_.max_seq_len * sizeof(int32_t), cudaHostAllocDefault) !=
+            cudaSuccess)
+            h_pf_token_ids_ = nullptr;
+    }
+
+    // Report memory
+    {
+        size_t free_mem = 0, total_mem = 0;
+        if (cudaMemGetInfo(&free_mem, &total_mem) == cudaSuccess)
+            IMP_LOG_INFO("GPU memory: %.0f MiB used / %.0f MiB total (%.0f MiB free)",
+                         (total_mem - free_mem) / (1024.0 * 1024.0), total_mem / (1024.0 * 1024.0),
+                         free_mem / (1024.0 * 1024.0));
+        memory_manager_.vram_allocator().report();
+    }
+
+    return true;
+}
+
 bool Engine::init_features() {
     const auto& mcfg = model_->config();
